feat: Add signout functionality
Added a new function `signout` to the `user.controller.js` file that clears the access token cookie and returns a successful signout message. The function is called when the `/users/signout` route is accessed via a POST request. If an error occurs during the signout process, a 500 status code is returned along with an error message.

This commit also includes updates to the `request.rest` file to include a sample request for the signout route.

Lastly, the `user.js` route file is updated to include the `/signout` route, which maps to the `signout` function in the `user.controller.js` file.